requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
requests
psutil
pickle-mixin
fastapi
uvicorn[standard]
python-dotenv
psutil
sentence-transformers
faiss-cpu
numpy
pandas
requests
pydantic